requirements.txt
amqp==2.5.1
backports.csv==1.0.7
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.7
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-mdeditor==0.1.16
django-mptt==0.10.0
django-pure-pagination==0.3.0
django-redis==4.10.0
django-reversion==3.0.4
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.8
importlib-metadata==0.23
jdcal==1.4.1
jieba==0.39
kombu==4.6.4
Markdown==3.1.1
more-itertools==7.2.0
odfpy==1.4.0
openpyxl==3.0.0
Pillow==6.1.0
Pygments==2.4.2
PyMySQL==0.9.3
pytz==2019.2
PyYAML==5.1.2
redis==3.3.8
requests==2.22.0
six==1.10.0
tablib==0.13.0
urllib3==1.25.6
uWSGI==2.0.18
vine==1.3.0
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
